requirements.txt
flask>=3.0.0
google-genai>=1.0.0
Pillow>=10.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
